# docs/admin/swagger.yaml
basePath: /admin/expense
definitions:
  ptime.TimeResponse:
    properties:
      time:
        type: string
    type: object
  requestmodel.CategoryBodyCreate:
    properties:
      name:
        type: string
      type:
        type: string
    type: object
  requestmodel.CategoryBodyUpdate:
    properties:
      name:
        type: string
    type: object
  requestmodel.CategoryChangeStatus:
    properties:
      status:
        type: string
    type: object
  requestmodel.StaffBodyChangePassword:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    type: object
  requestmodel.StaffBodyLogin:
    properties:
      password:
        type: string
      phone:
        type: string
    type: object
  requestmodel.StaffBodyUpdate:
    properties:
      email:
        type: string
      gender:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  requestmodel.SubCategoryBodyCreate:
    properties:
      name:
        type: string
    type: object
  requestmodel.SubCategoryChangeStatus:
    properties:
      status:
        type: string
    type: object
  responsemodel.ResponseCategoryAdmin:
    properties:
      _id:
        type: string
      createdAt:
        $ref: '#/definitions/ptime.TimeResponse'
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updatedAt:
        $ref: '#/definitions/ptime.TimeResponse'
    type: object
  responsemodel.ResponseCategoryAll:
    properties:
      limit:
        type: integer
      list:
        items:
          $ref: '#/definitions/responsemodel.ResponseCategoryAdmin'
        type: array
      total:
        type: integer
    type: object
  responsemodel.ResponseChangeStatus:
    properties:
      _id:
        type: string
      status:
        type: string
    type: object
  responsemodel.ResponseCreate:
    properties:
      _id:
        type: string
    type: object
  responsemodel.ResponseSubCategoryAdmin:
    properties:
      _id:
        type: string
      createdAt:
        $ref: '#/definitions/ptime.TimeResponse'
      name:
        type: string
      status:
        type: string
      type:
        type: string
      updatedAt:
        $ref: '#/definitions/ptime.TimeResponse'
    type: object
  responsemodel.ResponseSubCategoryAll:
    properties:
      limit:
        type: integer
      list:
        items:
          $ref: '#/definitions/responsemodel.ResponseSubCategoryAdmin'
        type: array
      total:
        type: integer
    type: object
  responsemodel.ResponseUpdate:
    properties:
      _id:
        type: string
    type: object
info:
  contact:
    email: tuanngo.472000@gmail.com
    name: Dev team
  description: |
    All APIs for Expense admin.

    ******************************
    - Add description
    ******************************
  title: Expense Tracker - Admin API
  version: "1.0"
paths:
  /categories:
    get:
      consumes:
      - application/json
      operationId: category-all
      parameters:
      - in: query
        name: keyword
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: status
        type: string
      - in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseCategoryAll'
      security:
      - ApiKeyAuth: []
      summary: All
      tags:
      - Category
    post:
      consumes:
      - application/json
      operationId: category-create
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.CategoryBodyCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseCreate'
      security:
      - ApiKeyAuth: []
      summary: Create
      tags:
      - Category
  /categories/{id}:
    get:
      consumes:
      - application/json
      operationId: category-detail
      parameters:
      - description: Category id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseCategoryAdmin'
      security:
      - ApiKeyAuth: []
      summary: Detail
      tags:
      - Category
    put:
      consumes:
      - application/json
      operationId: category-update
      parameters:
      - description: Category id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.CategoryBodyUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseUpdate'
      security:
      - ApiKeyAuth: []
      summary: Update
      tags:
      - Category
  /categories/{id}/status:
    patch:
      consumes:
      - application/json
      operationId: category-change-status
      parameters:
      - description: Category id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.CategoryChangeStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseChangeStatus'
      security:
      - ApiKeyAuth: []
      summary: ChangeStatus
      tags:
      - Category
  /categories/{id}/sub-categories:
    get:
      consumes:
      - application/json
      operationId: sub-category-all
      parameters:
      - description: Category id
        in: path
        name: id
        required: true
        type: string
      - in: query
        name: keyword
        type: string
      - in: query
        name: limit
        type: integer
      - in: query
        name: page
        type: integer
      - in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseSubCategoryAll'
      security:
      - ApiKeyAuth: []
      summary: Get all Sub Category
      tags:
      - Category
    post:
      consumes:
      - application/json
      operationId: sub-category-create
      parameters:
      - description: Category id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.SubCategoryBodyCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseCreate'
      security:
      - ApiKeyAuth: []
      summary: Create Sub Category
      tags:
      - Category
  /ping:
    get:
      consumes:
      - application/json
      operationId: ping
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Ping
      tags:
      - Common
  /staffs/login:
    post:
      consumes:
      - application/json
      operationId: staff-login
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.StaffBodyLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Login
      tags:
      - Staff
  /staffs/me:
    get:
      consumes:
      - application/json
      operationId: staff-get-me
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: GetMe
      tags:
      - Staff
    put:
      consumes:
      - application/json
      operationId: staff-update-info
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.StaffBodyUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update
      tags:
      - Staff
  /staffs/me/password:
    patch:
      consumes:
      - application/json
      operationId: staff-change-password
      parameters:
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.StaffBodyChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: ChangePassword
      tags:
      - Staff
  /sub-categories/{id}:
    get:
      consumes:
      - application/json
      operationId: sub-category-detail
      parameters:
      - description: Sub category id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseSubCategoryAdmin'
      security:
      - ApiKeyAuth: []
      summary: Detail
      tags:
      - SubCategory
    put:
      consumes:
      - application/json
      operationId: sub-category-update
      parameters:
      - description: Sub category id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.SubCategoryBodyCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseCreate'
      security:
      - ApiKeyAuth: []
      summary: Update
      tags:
      - SubCategory
  /sub-categories/{id}/status:
    patch:
      consumes:
      - application/json
      operationId: sub-category-change-status
      parameters:
      - description: Sub category id
        in: path
        name: id
        required: true
        type: string
      - description: Payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/requestmodel.SubCategoryChangeStatus'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responsemodel.ResponseChangeStatus'
      security:
      - ApiKeyAuth: []
      summary: ChangeStatus
      tags:
      - SubCategory
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"